% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireworks-dependencies.R
\name{add_fireworks_deps}
\alias{add_fireworks_deps}
\title{fireworks dependencies utils}
\usage{
add_fireworks_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches fireworks dependencies to the given tag
}
