% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireworks.R
\name{boom}
\alias{boom}
\title{Boom}
\usage{
boom()
}
\value{
The symbol \code{R_UnboundValue}.
}
\description{
Errors if attempting to print or assign the return value. Resistant to
    handling with \code{tryCatch()}.
}
\examples{
isS4(sparkle(boom()))

}
