% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireworks.R
\docType{package}
\encoding{UTF-8}
\name{fireworks-package}
\alias{fireworks}
\alias{fireworks-package}
\title{fireworks: Fireworks for R}
\description{
A toolbox for hacking on the language.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shikokuchuo.net/fireworks/}
  \item \url{https://github.com/shikokuchuo/fireworks/}
  \item Report bugs at \url{https://github.com/shikokuchuo/fireworks/issues}
}

}
\author{
Charlie Gao \email{charlie.gao@shikokuchuo.net}
    (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
}
