% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireworks.R
\name{fireworks}
\alias{fireworks}
\title{Add fireworks div}
\usage{
fireworks(id, ...)
}
\arguments{
\item{id}{The id of the fireworks div}

\item{options}{A list of options to pass to the fireworks}
}
\description{
Add fireworks div
}
