% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireworks.R
\name{fireworksClass}
\alias{fireworksClass}
\alias{Fireworks}
\title{Fireworks R6 Class}
\description{
Create fireworks and start or stop the fireworks.
}
\details{
Create an object to show fireworks as an overlay on the app
or a specific HTML element specified by \code{id}. Then \code{start},
the fireworks show, and \code{stop} the fireworks show.
}
\examples{

## ------------------------------------------------
## Method `Fireworks$new`
## ------------------------------------------------

\dontrun{Fireworks$new()}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fireworks-new}{\code{Fireworks$new()}}
\item \href{#method-fireworks-start}{\code{Fireworks$start()}}
\item \href{#method-fireworks-stop}{\code{Fireworks$stop()}}
\item \href{#method-fireworks-clone}{\code{Fireworks$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fireworks-new"></a>}}
\if{latex}{\out{\hypertarget{method-fireworks-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireworks$new(id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id, or vector of ids, of element on which to display the fireworks, if \code{NULL} the fireworks show
will be an overlay on the full HTML document.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create fireworks.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Fireworks$new()}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fireworks-start"></a>}}
\if{latex}{\out{\hypertarget{method-fireworks-start}{}}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireworks$start()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Start fireworks.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fireworks-stop"></a>}}
\if{latex}{\out{\hypertarget{method-fireworks-stop}{}}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireworks$stop()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Stop the fireworks
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fireworks-clone"></a>}}
\if{latex}{\out{\hypertarget{method-fireworks-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireworks$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
